requirements.txt
GitPython==3.1.29
matplotlib==3.5.2
nltk==3.7
numpy==1.21.5
pandas==1.4.2
requests==2.28.1
scipy==1.7.3
selenium==4.6.1
